refactor models for test passing

+++ main/models/tag.py
@@ -2,5 +2,5 @@ from django.db import models
 
 
 class Tag(models.Model):
-    id = models.PositiveIntegerField(primary_key=True, null=False, unique=True)
+    id = models.BigAutoField(primary_key=True, null=False, unique=True)
     title = models.CharField(max_length=255)
